int06-11

+++ workspace/ims-webapp/src/main/java/th/co/baiwa/excise/ia/persistence/repository/HealthCareWelFareRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.co.baiwa.excise.ia.persistence.entity.HealthCareWelFareEntity;
 
-public interface HealthCareWelFareRepository extends CommonJpaCrudRepository<HealthCareWelFareEntity, Long> {
+public interface HealthCareWelFareRepository extends CommonJpaCrudRepository<HealthCareWelFareEntity, BigDecimal> {
 
 	List<HealthCareWelFareEntity> findByStatusCheckAndIaDisReqIdIsNull(String string);
